⚔️ Definitely errors: communitacting errors
Implemented EIP-1193 Standard error handling and let it bubble to the dApps.

Also implemented DAppRequest deny on closing the request window.

Restructured `provider-bridge-shared` packages so it does not have dependency cycles
