+++ background/services/provider-bridge/index.ts
@@ -8,7 +8,7 @@ import {
   PortResponseEvent,
   EIP1193Error,
   RPCRequest,
-  EIP1193_ERROR,
+  EIP1193_ERROR_CODES,
   isTallyInternalCommunication,
 } from "@tallyho/provider-bridge-shared"
 import BaseService from "../base"
@@ -144,11 +144,13 @@ export default class ProviderBridgeService extends BaseService<Events> {
       } else {
         // if user does NOT agree, then reject
 
-        response.result = new EIP1193Error(EIP1193_ERROR.userRejectedRequest)
+        response.result = new EIP1193Error(
+          EIP1193_ERROR_CODES.userRejectedRequest
+        )
       }
     } else {
       // sorry dear dApp, there is no love for you here
-      response.result = new EIP1193Error(EIP1193_ERROR.unauthorized)
+      response.result = new EIP1193Error(EIP1193_ERROR_CODES.unauthorized)
     }
 
     port.postMessage(response)
