+++ provider-bridge-shared/runtime-typechecks.ts
@@ -1,0 +1,82 @@
+import {
+  RPCRequest,
+  WindowResponseEvent,
+  PortResponseEvent,
+  TallyInternalCommunication,
+} from "./types"
+
+export function getType(arg: unknown) {
+  return Object.prototype.toString.call(arg).slice("[object ".length, -1)
+}
+
+export function isObject(
+  arg: unknown
+): arg is Record<string | number | symbol, unknown> {
+  return getType(arg) === "Object"
+}
+
+export function isArray(arg: unknown): arg is Array<unknown> {
+  return Array.isArray(arg)
+}
+
+export function isUndefined(arg: unknown): arg is undefined {
+  return typeof arg === "undefined"
+}
+
+export function isString(arg: unknown): arg is string {
+  return getType(arg) === "String"
+}
+
+export function isNumber(arg: unknown): arg is number {
+  return getType(arg) === "Number"
+}
+
+export function isMessageEvent(arg: unknown): arg is MessageEvent {
+  return arg instanceof MessageEvent
+}
+
+export function isRPCRequestParamsType(
+  arg: unknown
+): arg is RPCRequest["params"] {
+  return isObject(arg) || isArray(arg)
+}
+
+export function isWindowResponseEvent(
+  arg: unknown
+): arg is WindowResponseEvent {
+  return (
+    isMessageEvent(arg) &&
+    isString(arg.origin) &&
+    !isUndefined(arg.source) &&
+    isObject(arg.data) &&
+    isString(arg.data.id) &&
+    isString(arg.data.target) &&
+    !isUndefined(arg.data.result)
+  )
+}
+
+export function isPortResponseEvent(arg: unknown): arg is PortResponseEvent {
+  return isObject(arg) && isString(arg.id) && !isUndefined(arg.result)
+}
+
+export const AllowedQueryParamPage = {
+  signTransaction: "/signTransaction",
+  dappPermission: "/dapp-permission",
+} as const
+
+export type AllowedQueryParamPageType =
+  typeof AllowedQueryParamPage[keyof typeof AllowedQueryParamPage]
+
+export function isAllowedQueryParamPage(
+  url: unknown
+): url is AllowedQueryParamPageType {
+  // The typing for Array.includes in `lib.es.2016.array.include.ts` does not make any sense here -> Object.values<string>
+  // interface Array<T> { ... includes(searchElement: T, fromIndex?: number): boolean; ...
+  return Object.values<unknown>(AllowedQueryParamPage).includes(url)
+}
+
+export function isTallyInternalCommunication(
+  arg: unknown
+): arg is TallyInternalCommunication {
+  return isObject(arg) && arg.method === "tally_getConfig"
+}
